Inventory system started
A basic inventory system is in place. Pressing E while pointing to an item pickup will pick it up if the extra wieght does not exceed capacity.
Place item pickups in the world and set the item it provides and the count. The mesh will not show in editor but will update when the game starts.
Create an item from assetmenu. Make sure to give them an unique id as that is what defines it internally.
The item manager does nothing as of now. It holds information about all items and can generate items during runtime.